first part works

+++ bamazon.sql
@@ -3,7 +3,7 @@ CREATE DATABASE bamazon;
 USE bamazon;
 
 CREATE TABLE products(
-    item_id  INTEGER(5) ZEROFILL NOT NULL,
+    item_id  INTEGER(4) ZEROFILL NOT NULL,
     product_name VARCHAR(128) NOT NULL,
     department_name VARCHAR(50) NOT NULL,
     price DECIMAL(10,2) NOT NULL,
